+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Create Execution component for displaying configurations and results [#5](https://github.com/IN-CORE/incore-studio/issues/5)
 - Ability to fill out parameters and input datasets for execution [#61](https://github.com/IN-CORE/incore-studio/issues/61)
 - Display status of each execution steps [#39](https://github.com/IN-CORE/incore-studio/issues/39)
+- Modal to create a new workflow and add to project [#41](https://github.com/IN-CORE/incore-studio/issues/41)
 
 ### Changed